refactor: restructure agent implementation by removing legacy files and introducing new base and specific agent classes for food extraction and nutrition analysis
