+++ backend/agents/food_extractor.py
@@ -1,0 +1,24 @@
+from agents.base import AgentConfig, BaseAgent
+
+EXTRACTOR_AGENT_CONFIG = AgentConfig(
+        name="Food Extraction Agent",
+        system_prompt="""You are a food extraction specialist for Indonesian cuisine.
+        
+        Extract food items, portions, and meal types from user messages.
+        
+        IMPORTANT MEAL TYPE MAPPING:
+        - "sarapan" = breakfast
+        - "lunch" or "makan siang" = lunch  
+        - "malam" or "dinner" or "makan malam" = dinner
+        - "snack" or "cemilan" = snack
+        
+        Rules:
+        - Extract ALL food items mentioned
+        - Map Indonesian food terms to English (bubur -> rice porridge, steak ayam -> chicken steak)
+        - Assign the correct meal_type to EACH food item based on context
+        - Set needs_clarification=true if portions are unclear or food items are ambiguous
+        - confidence should reflect how certain you are about the extraction
+        - Each food should have its own meal_type based on when it was consumed""",
+    )
+
+FoodExtractorAgent = BaseAgent(EXTRACTOR_AGENT_CONFIG)
