+++ backend/models/__init__.py
@@ -1,0 +1,2 @@
+"""Domain models for the application"""
+
