+++ backend/repositories/__init__.py
@@ -1,0 +1,2 @@
+"""Repository layer for data access"""
+
