+++ backend/repositories/models/__init__.py
@@ -1,0 +1,2 @@
+"""Repository-specific models and schemas"""
+
